+++ verifytrusted.php
@@ -3,10 +3,12 @@
  * Plugin Name:       Verify Trusted Reviews
  * Plugin URI:        https://www.verifytrusted.com/
  * Description:       Display aggregated reviews from Verify Trusted on your site.
- * Version:           1.3.0
+ * Version:           1.4.0
+ * Requires at least: 6.0
+ * Requires PHP:      8.1
  * Author:            Verify Trusted
  * License:           GPLv2 or later
- * License URI:       http://www.gnu.org/licenses/gpl-2.0.html
+ * License URI:       https://www.gnu.org/licenses/gpl-2.0.html
  * Text Domain:       verifytrusted
  * Domain Path:       /languages
  *
@@ -15,10 +17,11 @@
 
 defined( 'ABSPATH' ) || die();
 
-const VTRUST_VERSION = '1.3.0';
+const VTRUST_VERSION = '1.4.0';
 
 define( 'VTRUST_DIR', plugin_dir_path( __FILE__ ) );
 define( 'VTRUST_URL', plugin_dir_url( __FILE__ ) );
+define( 'VTRUST_BASENAME', plugin_basename( __FILE__ ) );
 
 require_once VTRUST_DIR . 'constants.php';
 require_once VTRUST_DIR . 'functions.php';
@@ -26,6 +29,7 @@ require_once VTRUST_DIR . 'functions.php';
 require_once VTRUST_DIR . 'includes/class-api-client.php';
 require_once VTRUST_DIR . 'includes/class-settings.php';
 require_once VTRUST_DIR . 'includes/class-admin-hooks.php';
+require_once VTRUST_DIR . 'includes/class-shortcode.php';
 require_once VTRUST_DIR . 'includes/class-plugin.php';
 
 /**